Trinity.C: network: unifying epoll/kqueue implementation a bit
